Renames package to j8spec.
