+++ src/main/java/j8spec/MyObject.java
@@ -1,4 +1,4 @@
-package specs;
+package j8spec;
 
 public class MyObject {
     private int value;
