+++ src/main/java/j8spec/Specs.java
@@ -1,4 +1,4 @@
-package specs;
+package j8spec;
 
 import java.util.*;
