+++ src/main/java/j8spec/Var.java
@@ -1,0 +1,5 @@
+package j8spec;
+
+public class Var<T> {
+    public T value = null;
+}
