--- src/main/java/specs/Var.java
@@ -1,5 +1,0 @@
-package specs;
-
-public class Var<T> {
-        public T value = null;
-}
